More changed to meta
I added animation to the opening… although it looks cool, I think it’s
a bit much for the browser to handle, so I’m going to have to change it
I think… maybe change it to a kind of bubble expand/contract effect
instead. I also went back to the darkgrey color… I have decided I like
this.